Move keyboard handling to window so we can do line editing
The logic for line editing is a little complicated, so it's
best not to duplicate it across all the plugins.  Now we
manage it all from the window.  The plugins now access the
various editing events via there vtable, but that's an
extra layer of indirection that doesn't matter given that
we pass the window to the plugins anyway.  We should drop
that and just have the plugins register for edit events
directly.