Fix bug in area union code.
ply_frame_buffer_add_area_to_flush_area() computes the wrong area
when the new area right of the current area but is wider than the
old area.

The easiest way to get rectangle intersection right is to just do
it on absolute coordinates and then convert back to width and height
afterwards.