Compute correct offset when copying from shadow buffer to frame buffer
When I added commit c12164c1622a2209fe07555e682cc479c6854e7e
to copy an entire row at a time to the framebuffer, I miscalculated the
offset to copy from/to.  Their may be uninitialized data at the
beginning and end of the temporary row buffer, if only part of the row
is getting copied.  We need to make sure we jump passed that junk in
memory and copy just the part that got filled in.