Use more descriptive format for control code #defines
The previous format was just the ascii value directly
in octal.  I'm getting sick of typing 'man ascii' to look
it up though.  The new format is just, e.g.,

('\100' ^ 'U')

which not only means I can add new entries without looking
them up, but also has the added advantage that it looks
sort of similiar to, e.g., ^U, which is a common notation
for respresenting control characters.