Add ctrl-u and ctrl-w to erase password line
Right now we do it in the cheesiest way possible, by
calling the backspace function over and over again on
behalf of the user.  It might make more sense to export
another window callback specifically for erase line. It
probably doesn't make sense to do that until we fix the
TODO item:

- have plugins hook into line editing through window
  directly, instead of via vtable functions

though.