add new answer object for giving a deferred response
Right now we have this hack tower of function pointers
cutting through layers of code to get the password
from the splash plugin to the client.  I'd like to cut
most of that out by just having one answer object that
the bottom layer and the the top layer know directly,
without the middle layers having to get too involved.