Generate background instead of using image
The background image previously was low resolution
and created various artifacts when being upscaled
to the native resolution of the panel.  This patch
ditches the background image and instead generates
a similar type of background dynamically, custom
fit to the screen it's being displayed on.  As an
added bonus, since we're doing it dynamically, we
can make some of the stars in the background
lightly twinkle.