[libply] Add new keyfile class
This is a very lame keyfile parser.  We'll be able to use it to load
some metadata about the plugins.

The idea is that plugins will shift to being engines that can be
leveraged from multiple splash themes.  This way, for instance, the glow
plugin (which will be renamed) can be used to implement several
splashes.