[frame-buffer] Track multiple flush areas
Previously we would always aggregate flush areas together even
if they were disjoint and far apart.  That meant that if two
images on opposites sides of the screen were updated in one
frame, then the entire screen would get redrawn.

We now track flush areas in a list, instead of a bounding box.