[window] Handle disconnected tty fd better
Right now if we ever get a hangup from the mainloop
on the tty fd then the window's fd watch on the tty
becomes invalid and we will ultimately crash when
trying to disconnect the watch later.

This commit sets the watch to NULL if a hang up happens,
and adds some debug spew.