[libply] Add new region class
Sometimes it's necessary to deal with a disjoint
set of rectangles together.

One big example is when keeping track of which parts of
the screen have changed.

Another example would be for non-rectangular clip masks.

This commit adds a new region class to simplify the handling
of these types of cases.