[libplybootsplash] Add new pixel buffer class
Right now a lot of pixel manipulation functions are in the frame buffer
code.  This means it's hard to do these types of things without the
results showing up on the display.

Furthermore, we want to get away from one frame buffer anyway for
multihead support.  I can imagine each display in a multihead
configuration with its own pixel buffer, so splash plugins can draw to
the independently.

This commit adds the start of a pixel buffer class, which just copies
and pastes most of pixel-twiddling frame buffer code.  There are some
changes to deal with pixel buffers not always having a final alpha of
1.0, and some changes to provide a more flexible interface for masking
fill operations.

At some point it will probably gain copy-and-pastes of the resize an
rotate code from the image class as well, and provide a mechansim for
have subpixbufs where rowstride != width.