[utils] Force modules to stay resident after close
Some modules (the label plugin in particular) use
libraries with static data that don't do well with
being unmapped and remapped later.

This commit changes the module loading call so that
plugins aren't unmapped at close time (the close
is effectively ignored).  This simplifies things
for now.

At some point we may want to make it decidable on
a per plugin basis.