[event-loop] Add ply_event_loop_get_default ()
The event loop is used all through out the code
for many different things.  In order to use it
in all those various bits of code, it's getting
passed around a lot.  This adds a lot of complexity
to the code.

The event loop should probably be a standard
resource that all parts of the code have free access
to in the same vein as the logging apis.

This commit adds a new function
ply_event_loop_get_default () and makes main use it
and pass it along.

Subsequent commits may clean up all the places that
take an event loop to just use this instead.