[scripts] add plymouth-generate-initrd script
One useful way to use plymouth is by installing it into
an auxillary initrd image, that overlays the primary one.

This allows plymouth and its current theme to get updated
indepedently of the kernel and the rest of the initrd
stuff.

plymouth-generate-initrd creates an initrd named:

     initrd-plymouth.img

in /boot that is suitable for use a second initrd on
the initrd line in grub.conf.