[main] deactivate renderer before hiding
If we don't deactivate the renderer before hiding the splash, the
drm renderer may scan out the buffer contents to the fbcon buffer;
since we only hide the splash when dumping details or when
--retain-splash is *not* given to quit, this is exactly the
opposite of what we want.

The effect of not doing this is partial splash contents behind the
details in cases of error, or when using quit.  This doesn't affect
plymouth quit --retain-splash.