[script] Get window value without an index returns area covered by all windows
Calls to Window.GetWidth/Height/X/Y without a window index now return the
values of the area covered by all windows. This is only the case if all the
windows are aligned (either by their centers, or to a corner).

This allows the theme designer to place an object knowing it will be seen on
all screens.