[main] Try getting default theme from config file
Right now we figure out the default theme via a symlink.
This apprach is very simple, but is also a little cumbersome.

It means as the default theme is changed around we have to move
the symlink around.

The symlink is in /usr.  We really shouldn't be mucking with
/usr when changing defaults.

This commit checks the filesystem for two config files:

/usr/share/plymouth/plymouthd.defaults
and
/etc/plymouth/plymouthd.conf

The first one is for distributions to use.  This is how they can
manage which splash to show from release to release.

The second one is for system administrators.  This is how they
can override distribution policy.

We don't actually ship these files yet.  In the mean time,
(and even after for the forseeable future) the old symlink method
will still work.