[main] watch keyboard input after displays are added
Since the boot splashes handle opening the terminal now,
we need to make sure that we don't try to watch for input on
the terminal until they've opened it.

This should fix some crashes users are reportedly seeing with
the details splash.

A potentially better fix would be to make watch_for_input defer
doing anything on its own using the terminal is opened.  That
would require a trigger and some new api to ply-terminal, I think.