[main] Watch keyboard input after the splash had a chance to quit
If the splash hasn't picked up any displays it will abort. We wait for the
possible abort before attaching the keyboard. This is a slight hack as now
add_displays_and_keyboard_to_boot_splash no longer adds the keyboards. Also the
splash plugin can access the keyboard before it is officially opened.

For now it just fixes the terminal fd < 0 assertion bug.