[main] Ignore keyboard changes when deactivating
There is a small window after plymouth is told
to quit or deactivate before it actually does.

During this window, if the user happens to hit
escape, bad things could happen.

We really don't want to mess with, e.g., the tty
settings when we're about to exit.

This commit puts plymouth in a sort of "degraded"
mode while it's deactivating.  During this time
frame, user input is ignored.