[main] Create a list of messages and when showing splash display them
The system now keeps a list of messages so they can be shown again when the
user flips themes or if the messages were passed before the splash was shown.

This also enables the splashes to show multiple messages (should they choose to
do so). There will later be a way of undisplaying a message.