[throbber] Change frame sequencer function
The throbber has a function for determining which frame
to show based on the current time.  This function was:

  ƒrame(t) = number_of_frames ∙ (⅟₂ sin(t) + ⅟₂)

Which basically oscillates between 0 and number_of_frames - 1,
over and over again.  There are two problems with this function;
  - after it runs through all the frames in order, it then procedes
    to run through them backward.
  - This function also starts in the middle of the set of throbber
    frames.

These problems don't matter for spinfinity, but will look wrong for most
other themes.

The new function is this:

  ƒrame(t) = number_of_frames ∙ (1⁄duration)(t mod duration)

This function solves both problems.  At time 0 it uses frame 0, and
after the last frame it jumps back to the first frame.