[main] Be more forgiving of an absent boot splash
Mandriva has cases where they need to drop to details mode
before /usr is mounted.  /usr has the details splash plugin,
so this causes plymouth to explicitly exit.

This commit makes plymouth cope with that scenario slightly
better.  It takes out the exit(1) call which was clearly bogus,
and also makes parts of the code handle a NULL splash better.
This is not a real fix though.

Ultimately, we should make "details" be built-in as a nice failsafe.
There's no reason we should need to load it off the disk as a plugin.