[drm] free stored connector
For most connectors associating monitors with our virtual
'heads', we merely track their ids.  This means we don't have
to bother freeing any client side state at deallocation time.

There's one exception, though.  The main connector, connector0,
we keep an open reference to.  We do this, because it owns the
mode object we use in SetCrtc calls.

This commit ensures that connector0 for each head is properly freed
when that head is deallocated.