terminal: unlock tty before mucking with it
It seems that sometimes when the tty is stolen from us
it remains locked when we get it back.  That would be okay
if the attributes were still kosher, but they aren't.

This commit unlocks the tty before trying to reinitialize
the terminal with suitable attributes.

Debugged by Andrey Borzenkov <arvidjaar@gmail.com>