terminal: Keep retrying to reopen tty on EIO after hang up
If the tty hangs up on us, the kernel will return EIO while it's
closing down.  As far as I know, there's no way to get notified
when it's back up again, so just keep retrying in a loop over
the course of a second or so.  Eventually give up, though.

This is really suboptimal and suggests we should get away from
using ttys for input at some point.