systemd: toggle systemd messages when toggling console redirection
Normally systemd is very mute about messages.  It's important,
though for it to be chatty when plymouth is running so we can
show verbose messages to the user when they hit escape.

This commit adds a new --enable-systemd-integration configure
flag which explicitly tells systemd when to print messages to
the console.

This may get dropped in the future in lieu of init script
changes doing this instead of plymouth directly.