keyboard: handle kernel terminal EIO better
In commit 89096d735f78fab300248962030d6554f9894011 we added
reopen retries when the kernel returns EIO to ply-terminal.c.

This is because when the kernel is closing a tty down, that
tty is unavailable to userspace to reopen.

Unfortunately, that commit neglected to inform the ply-keyboard
part of the code when the terminal retry was successful.  The
upshot of this, is that if plymouthd needs to retry opening the
tty, then the splash screens lose control over the keyboard.

This commit changes how input notification is sent to the keyboard
handling code, so the tty disconnects are transparent.