x11: translate enter to new line
commit 4081bd29fb471a93daba746643419d858f06bdfe changed
our keyboard handling code to treat newline instead of
carriage return as the enter key.

That commit fixed up the tty settings to work under these
new assumptions.  It failed to update the X11 code, though.

This commit fixes the X11 renderer plugin so it works in
the same way.